Add files via upload
In this project I learn how to build a Facebook Prophet Machine learning model in order to forecast the price of Tesla 30 days into the future. I also visualize the historical performance of Tesla through graphs and charts using Plotly express and evaluate the performance of the model against real data using Google Finance in Google Sheets. I  also dive into a brief stock analysis of Tesla and we will discuss PE ratio, EPS, Beta, Market cap, Volume and price range of Tesla. I end the project by automating the forecasting process in such a way that the forecast of any of other favourite stock with all necessary visualization within a few seconds of uploading the data. By the end of this project, I get confident in analyzing, visualizing and forecasting the price of any stock of choice.